去掉ILRuntime层的寻路IntPtr

+++ Unity/Codes/Model/Module/Recast/NavmeshComponent.cs
@@ -10,6 +10,6 @@ namespace ET
         // 为了区分客户端跟服务端还有机器人的加载方式，客户端是加载ab包，服务端跟机器人是加载bytes
         public Func<string, byte[]> Loader;
         
-        public Dictionary<string, IntPtr> Navmeshs = new Dictionary<string, IntPtr>();
+        public Dictionary<string, long> Navmeshs = new Dictionary<string, long>();
     }
 }
